Don't try to remove a policy that doesn't exist

+++ ansible/modules/hashivault/hashivault_policy_delete.py
@@ -63,6 +63,8 @@ from ansible.module_utils.hashivault import *
 def hashivault_policy_delete(params):
     name = params.get('name')
     client = hashivault_auth_client(params)
+    if name not in client.list_policies():
+        return {'changed': False}
     client.delete_policy(name)
     return {'changed': True}
